added in the settings cache
The settings cache was changed from db to be a dets/ets combination.

The dets is used for the persitence and crash restoration and the ets
is used for the lookups. I went this route because the lookup time of
ets v dets was 3x - 5x times as fast and since this will be getting
called on most pages I wanted to make sure it would be fast.